requirements.txt
langchain>=0.2.0
langchain-community>=0.0.26
langchain-openai>=0.0.8
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
